+++ java_console/io/src/main/java/com/rusefi/binaryprotocol/IncomingDataBuffer.java
@@ -45,14 +45,22 @@ public class IncomingDataBuffer {
     }
 
     public byte[] getPacket(String msg) throws EOFException {
-        return getPacket(msg, System.currentTimeMillis());
+        return getPacket(Timeouts.BINARY_IO_TIMEOUT, msg, System.currentTimeMillis());
+    }
+
+    public byte[] getPacket(String msg, long start) throws EOFException {
+        return getPacket(Timeouts.BINARY_IO_TIMEOUT, msg, start);
+    }
+
+    public byte[] getPacket(int timeoutMs, String msg) throws EOFException {
+        return getPacket(timeoutMs, msg, System.currentTimeMillis());
     }
 
     /**
      * why does this method return NULL in case of timeout?!
      * todo: there is a very similar BinaryProtocolServer#readPromisedBytes which throws exception in case of timeout
      */
-    public byte[] getPacket(String msg, long start) throws EOFException {
+    public byte[] getPacket(int timeoutMs, String msg, long start) throws EOFException {
         boolean isTimeout = waitForBytes(msg + " header", start, 2);
         if (isTimeout) {
             if (Bug3923.obscene)
@@ -66,7 +74,7 @@ public class IncomingDataBuffer {
         if (packetSize < 0)
             return null;
 
-        isTimeout = waitForBytes(loggingPrefix + msg + " body", start, packetSize + 4);
+        isTimeout = waitForBytes(timeoutMs, loggingPrefix + msg + " body", start, packetSize + 4);
         if (isTimeout)
             return null;