+++ java_console/ui/src/main/java/com/rusefi/SerialPortScanner.java
@@ -1,11 +1,16 @@
 package com.rusefi;
 
+import com.rusefi.autodetect.SerialAutoChecker;
+import com.rusefi.binaryprotocol.IncomingDataBuffer;
+import com.rusefi.io.IoStream;
 import com.rusefi.io.LinkManager;
+import com.rusefi.io.UpdateOperationCallbacks;
+import com.rusefi.io.serial.BufferedSerialIoStream;
 import com.rusefi.io.tcp.TcpConnector;
 import com.rusefi.maintenance.DfuFlasher;
-import com.rusefi.io.UpdateOperationCallbacks;
 import org.jetbrains.annotations.NotNull;
 
+import java.io.IOException;
 import java.util.*;
 import java.util.concurrent.CopyOnWriteArrayList;
 
@@ -83,6 +88,8 @@ public enum SerialPortScanner {
             startTimer();
     }
 
+    private static Map<String, PortResult> portCache = new HashMap<>();
+
     /**
      * Find all available serial ports and checks if simulator local TCP port is available
      */
@@ -92,43 +99,62 @@ public enum SerialPortScanner {
 
         String[] serialPorts = LinkManager.getCommPorts();
 
-        boolean hasAnyEcu = false;
+        int ecuCount = 0;
         boolean hasAnyOpenblt = false;
 
         for (String serialPort : serialPorts) {
-            // Filter out some macOS trash
-            if (serialPort.contains("wlan-debug") ||
-                    serialPort.contains("Bluetooth-Incoming-Port") ||
-                    serialPort.startsWith("cu.")) {
-                continue;
-            }
+            // First, check the port cache
+            if (portCache.containsKey(serialPort)) {
+                // We've already probed this port - don't re-probe it again
+                PortResult cached = portCache.get(serialPort);
 
-            // TODO: removeme once we can query each port's type
-            hasAnyEcu = true;
-            hasAnyOpenblt = true;
+                if (cached.type == SerialPortType.FomeEcu) {
+                    ecuCount++;
+                } else if (cached.type == SerialPortType.OpenBlt) {
+                    hasAnyOpenblt = true;
+                }
 
-
-            // See if this looks like an ECU
-            boolean isEcu = true;   // TODO
-            if (isEcu) {
-                ports.add(new PortResult(serialPort, SerialPortType.FomeEcu));
-                hasAnyEcu = true;
+                ports.add(cached);
             } else {
-                boolean isOpenblt = true;   // TODO
+                // This one isn't in the cache, probe it to determine what it is
+                PortResult result;
+
+                boolean isOpenblt = isPortOpenblt(serialPort);
                 if (isOpenblt) {
-                    ports.add(new PortResult(serialPort, SerialPortType.OpenBlt));
+                    result = new PortResult(serialPort, SerialPortType.OpenBlt);
                     hasAnyOpenblt = true;
                 } else {
-                    // Dunno what this is, leave it in the list anyway
-                    ports.add(new PortResult(serialPort, SerialPortType.Unknown));
+                    // See if this looks like an ECU
+                    boolean isEcu = isPortFomeEcu(serialPort);
+                    if (isEcu) {
+                        result = new PortResult(serialPort, SerialPortType.FomeEcu);
+                        ecuCount++;
+                    } else {
+                        // Dunno what this is, leave it in the list anyway
+                        result = new PortResult(serialPort, SerialPortType.Unknown);
+                    }
                 }
+
+                ports.add(result);
+                portCache.put(serialPort, result);
             }
         }
 
-        if (!ports.isEmpty()) {
-            ports.add(0, new PortResult(AUTO_SERIAL, SerialPortType.None));
+        // Clean the port cache of any entries that no longer exist
+        // If the same port appears later, we want to re-probe it at that time
+        // In any other scenario, auto could have unexpected behavior for the user
+        for (String x : portCache.keySet()) {
+            if (Arrays.stream(serialPorts).noneMatch(x::equals)) {
+                portCache.remove(x);
+            }
         }
 
+        boolean hasAnyEcu = ecuCount > 0;
+
+        // Only offer auto if there is more than one option to pick from, AND exactly one option is an ECU
+        if (ports.size() > 1 && ecuCount == 1) {
+            ports.add(0, new PortResult(AUTO_SERIAL, SerialPortType.None));
+        }
 
         if (includeSlowLookup) {
             for (String tcpPort : TcpConnector.getAvailablePorts()) {
@@ -205,6 +231,55 @@ public enum SerialPortScanner {
 
         public boolean isEmpty() {
             return !dfuFound && ports.isEmpty();
+        }
+    }
+
+    public static boolean isPortFomeEcu(String port) {
+        try (IoStream stream = BufferedSerialIoStream.openPort(port)) {
+            String signature = SerialAutoChecker.checkResponse(stream, null);
+
+            if (signature == null) {
+                return false;
+            }
+
+            return signature.contains("FOME");
+        } catch (Exception e) {
+            return false;
+        }
+    }
+
+    public static boolean isPortOpenblt(String port) {
+        try (IoStream stream = BufferedSerialIoStream.openPort(port)) {
+            if (stream == null) {
+                return false;
+            }
+
+            byte[] request = new byte[3];
+            request[0] = 2; // packet length
+            request[1] = (byte) 0xff; // XCPLOADER_CMD_CONNECT
+            request[2] = 0; // connectMode
+
+            stream.write(request);
+
+            IncomingDataBuffer idb = stream.getDataBuffer();
+
+            byte responseLength = idb.readByte(50);
+
+            // Invalid length, ignore
+            if (responseLength != 8) {
+                return false;
+            }
+
+            // Read length worth of bytes
+            byte[] response = new byte[responseLength];
+            idb.waitForBytes(100, "openblt detect", System.currentTimeMillis(), responseLength);
+            idb.read(response);
+
+            // Response packet should start with FF
+            // Not much else to check, as the rest of the response is protocol settings from the device.
+            return response[0] == (byte) 0xFF;
+        } catch (IOException e) {
+            return false;
         }
     }
 }